pyproject.toml
[project]
name = "knowsee"
version = "0.1.0"
description = "A bot for displaying calendars"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "beartype>=0.21.0",
    "icalendar>=6.3.1",
    "icalendar-compatibility>=0.1.4",
    "python-telegram-bot>=22.3",
    "pyyaml>=6.0.2",
    "recurring-ical-events>=3.8.0",
    "urllib3>=2.5.0",
]

[tool.ruff]
line-length = 72
target-version = "py313"
extend-include = ["*.ipynb"]

[tool.ruff.lint]
typing-modules = ["beartype.typing"]
select = [
  "E4", "E7", "E9",
  "F",
  "UP", # pyupgrade
  "I",  # isort
]
ignore = ["F722", "E741"]

[tool.ruff.lint.isort]
force-wrap-aliases = true
combine-as-imports = true
